requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
alembic==1.16.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asteroid-filterbanks==0.4.0
attrs==25.3.0
audioread==3.0.1
av==14.4.0
babel==2.17.0
blinker==1.9.0
blis==1.3.0
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
confection==0.1.5
contourpy==1.3.2
cryptography==45.0.3
csvw==3.5.1
cucco==2.2.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dacite==1.9.2
dataclasses-json==0.6.7
datasets==3.6.0
dateparser==1.2.2
decorator==5.2.1
dill==0.3.8
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
einops==0.8.1
einx==0.3.0
espeakng-loader==0.2.4
faiss-cpu==1.11.0
fastapi==0.115.12
fastrtc==0.0.26
fastrtc-moonshine-onnx==20241016
ffmpy==0.6.0
filelock==3.18.0
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.2
frozendict==2.4.6
frozenlist==1.6.2
fsspec==2025.3.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.41.0
google-crc32c==1.7.1
googleapis-common-protos==1.70.0
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
humanfriendly==10.0
HyperPyYAML==1.2.2
idna==3.10
ifaddr==0.2.0
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
julius==0.2.7
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.76
langchain-google-genai==2.1.12
langchain-anthropic
langchain-ollama==0.3.3
langchain-openai==0.3.33
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.11.0
lightning==2.5.1.post0
lightning-utilities==0.14.3
llvmlite==0.44.0
loguru==0.7.3
Mako==1.3.10
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
misaki==0.9.4
mlx==0.26.1
mlx-audio==0.2.3
mlx-lm==0.25.0
mlx-vlm==0.1.26
more-itertools==10.7.0
mpmath==1.3.0
msal==1.32.3
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
narwhals==1.42.1
networkx==3.5
num2words==0.5.14
numba==0.61.2
numpy==2.2.6
ollama==0.5.1
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.109.0
openai-whisper==20240930
opencv-python==4.10.0.84
optree==0.16.0
optuna==4.3.0
orjson==3.10.18
packaging==24.2
pandas==2.2.3
phonemizer-fork==3.3.2
pillow==11.2.1
platformdirs==4.3.8
pooch==1.8.2
preshed==3.0.10
primePy==1.3
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.1
psutil==5.9.8
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pyloudnorm==0.1.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytorch-lightning==2.5.1.post0
pytorch-metric-learning==2.8.1
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
ruff==0.11.13
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
segments==2.3.0
semantic-version==2.10.0
semver==3.0.4
sentence-transformers==5.1.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
speechbrain==1.0.3
SQLAlchemy==2.0.41
srsly==2.5.1
starlette==0.46.2
streamlit==1.45.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboardX==2.6.2.2
thefuzz==0.22.1
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.13.3
torch==2.7.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.7.1
torchmetrics==1.7.2
tornado==6.5.1
tqdm==4.67.1
transformers==4.52.4
typeguard==4.4.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.4.0
uvicorn==0.34.3
wasabi==1.1.3
watchdog==4.0.2
weasel==0.4.1
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
wmill